Properly detect z_v2settling_probe extension include order
The z_v2settling_probe.py extension work correctly only if it's included
in the config after the '[probe]' section. Otherwise, there is no good
way to avoid command and object conflicts.

This commit includes changes to the z_v2settling_probe extension that will
detect if the 'probe' object does not yet exist (which normally means
that the config parse has not processed the section yet). If the 'probe'
section does not exist, generate an error with an appropriate message.